+++ Userland/Services/SQLServer/ConnectionFromClient.h
@@ -29,9 +29,9 @@ private:
     explicit ConnectionFromClient(NonnullOwnPtr<Core::Stream::LocalSocket>, int client_id);
 
     virtual Messages::SQLServer::ConnectResponse connect(DeprecatedString const&) override;
-    virtual Messages::SQLServer::PrepareStatementResponse prepare_statement(int, DeprecatedString const&) override;
-    virtual void execute_statement(int, Vector<SQL::Value> const& placeholder_values) override;
-    virtual void disconnect(int) override;
+    virtual Messages::SQLServer::PrepareStatementResponse prepare_statement(u64, DeprecatedString const&) override;
+    virtual void execute_statement(u64, Vector<SQL::Value> const& placeholder_values) override;
+    virtual void disconnect(u64) override;
 };
 
 }
